An adjustment of class time_encode_too()
Adjusted the position of class time_encode_too().
