+++ demos/P300_demo.py
@@ -7,7 +7,7 @@ from metabci.brainda.algorithms.utils.model_selection import (
     generate_char_indices,
     match_char_kfold_indices
 )
-from metabci.brainda.utils.time_encode_tool import (
+from metabci.brainda.algorithms.decomposition.base import (
     TimeDecodeTool
 )
 from metabci.brainda.algorithms.decomposition import DCPM
